+++ test/fixtures/Parser/Client/Browser.php
@@ -8,6 +8,11 @@ class Browser
         'IE' => 'Internet Explorer'
     );
 
+    protected static $browserFamilies = array(
+        'Chrome'            => array('CH', 'CM'),
+        'Internet Explorer' => array('IE')
+    );
+
     protected static $mobileOnlyBrowsers = array(
         'CH', 'FF', 'IE'
     );